requirements.txt
tensorflow>=2.15.0
tensorflow-datasets>=4.9.4
streamlit>=1.32.0
pillow>=10.2.0
scikit-learn>=1.4.1
matplotlib>=3.8.3
seaborn>=0.13.2
pandas>=2.2.1
numpy>=1.26.4
